+++ src/core/api/SimBriefAeroflyApi.ts
@@ -68,7 +68,7 @@ export class SimBriefAeroflyApi extends SimBriefApi {
 
     const waypoints = this.getWaypointsFromNavlog(simbriefPayload);
     flight.navigation = new AeroflyNavigationConfig(
-      waypoints.reduce((acc, wp) => Math.max(acc + (wp.altitude ?? 0)), 0), // max altitude of all waypoints for cruise altitude
+      waypoints.reduce((acc, wp) => Math.max(acc , (wp.altitude ?? 0)), 0), // max altitude of all waypoints for cruise altitude
       [
         new AeroflyNavRouteOrigin(
           simbriefPayload.origin.icao_code,